Pull out Environment and beginning of Term
