+++ src/Abstract/Term.hs
@@ -1,0 +1,3 @@
+module Abstract.Term where
+
+type Name = String
